Simplify setup pass
